# KP-1yul Gas optimisation using standalone yul