Standalone YUL implementation of optimised gas contract.
GasContract is converted to an interface.
A contract factory function is used to deploy the YUL compiled code.
A one character change is made to setUp() in test/Gas.UnitTests.t.sol to call the factory function instead of the standard solidity constructor.

ALTHOUGH THE TEST CASES ARE UNMODIFIED, THE TEST SETUP FUNCTION REQUIRES A ONE CHARACTER DELETION AND IS THEREFORE INADMISSIBLE UNDER A STRICT INTERPRETATION OF THE NO CHANGE TO TEST CASE RULE.

However, implementing the entirety of the constract in YUL allows for a further 50% reduction in gas deployment cost over solidity and worth consideration.

Note throwing out significant amount of solidty generated YUL code breaks the deployment gas cost and size reporting of forge, therefore the contract factory function was instrumented to capture log gas consumed in calling CREATE and the size of the contract construction byte code passed to CREATE.

### Solidity only optimised gas reporting
```
+=======================================================================================+
| Deployment Cost                  | Deployment Size |       |        |       |         |
|----------------------------------+-----------------+-------+--------+-------+---------|
| 257431                           | 1566            |       |        |       |         |
|----------------------------------+-----------------+-------+--------+-------+---------|
```

### Solidity only optimised implemention with instrument contract deployment
```
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 178513, ~: 178604)
Logs:
  CREATE gas used & payload size: 289518 1566

Suite result: ok. 12 passed; 0 failed; 0 skipped; finished in 482.53ms (2.10s CPU time)

╭--------------------------------------+-----------------+-------+--------+-------+---------╮
| src/Gas.sol:GasContractImpl Contract |                 |       |        |       |         |
+===========================================================================================+
| Deployment Cost                      | Deployment Size |       |        |       |         |
|--------------------------------------+-----------------+-------+--------+-------+---------|
| 257407                               | 1566            |       |        |       |         |
|--------------------------------------+-----------------+-------+--------+-------+---------|
```

### Best YUL implmentation that doesn't break forge gas reporting
```
[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 178234, ~: 178460)
Logs:
  CREATE gas used & payload size: 270147 1476

Suite result: ok. 12 passed; 0 failed; 0 skipped; finished in 462.32ms (2.02s CPU time)

╭--------------------------------------+-----------------+-------+--------+-------+---------╮
| src/Gas.sol:GasContractImpl Contract |                 |       |        |       |         |
+===========================================================================================+
| Deployment Cost                      | Deployment Size |       |        |       |         |
|--------------------------------------+-----------------+-------+--------+-------+---------|
| 238040                               | 1476            |       |        |       |         |
|--------------------------------------+-----------------+-------+--------+-------+---------|
```

### Best YUL implementation (contract factory instrumentation numbers only)
```
% forge test --gas-report -vv
[⠊] Compiling...
[⠒] Compiling 2 files with Solc 0.8.28
[⠢] Compiling 1 files with Solc 0.8.28
[⠆] Solc 0.8.28 finished in 11.57ms
[⠰] Solc 0.8.28 finished in 1.06s
Compiler run successful!

Ran 12 tests for test/Gas.UnitTests.t.sol:GasTest
[PASS] testAddToWhitelist(uint256) (runs: 256, μ: 35761, ~: 35684)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] testBalanceOf() (gas: 14946)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] testCheckForAdmin() (gas: 11096)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] testTransfer(uint256,address) (runs: 256, μ: 87584, ~: 89395)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] testWhiteTranferAmountUpdate(uint256,string,uint256) (runs: 256, μ: 186587, ~: 186687)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] test_admins() (gas: 28736)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] test_onlyOwner(address,uint256) (runs: 256, μ: 35170, ~: 35277)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] test_tiers(address,uint256) (runs: 256, μ: 38417, ~: 38526)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] test_tiersReverts(address,uint256) (runs: 256, μ: 34647, ~: 34684)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] test_whitelistEvents(address,address,uint256,string,uint256) (runs: 256, μ: 177416, ~: 177503)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] test_whitelistEvents(address,uint256) (runs: 256, μ: 41454, ~: 41564)
Logs:
  CREATE gas used & payload size: 172362 628

[PASS] test_whitelistTransfer(address,address,uint256,string,uint256) (runs: 256, μ: 177406, ~: 177484)
Logs:
  CREATE gas used & payload size: 172362 628

Suite result: ok. 12 passed; 0 failed; 0 skipped; finished in 467.00ms (2.09s CPU time)


Ran 1 test suite in 521.31ms (467.00ms CPU time): 12 tests passed, 0 failed, 0 skipped (12 total tests)
```